extracted_repo/requirements.txt
aenum==3.1.16
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
attrs==25.3.0
beautifulsoup4==4.13.4
binance-connector==3.12.0
ccxt==4.5.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cryptography==45.0.6
curl_cffi==0.13.0
cycler==0.12.1
defusedxml==0.7.1
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.111.0
fastapi-cli==0.0.8
fonttools==4.59.1
frozendict==2.4.6
frozenlist==1.7.0
future==1.0.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpie==3.2.4
httptools==0.6.4
httpx==0.27.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.9
line-bot-sdk==3.18.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mplfinance==0.12.10b0
multidict==6.6.4
multitasking==0.0.12
numpy==2.3.2
openpyxl==3.1.5
orjson==3.11.2
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
protobuf==6.32.0
pycares==4.10.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.8.2
pydantic-settings==2.10.1
pydantic_core==2.20.1
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.0
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.37.2
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.30.3
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
yarl==1.20.1
yfinance==0.2.65
